Computes mixture normal variance

Input the desired variance, number of distributions, and mean of the distributions, returns a value of the variance of each mixture distribution.

desireVar(desVar, num_dist, means, equalWeight = TRUE)

Arguments

desVar
Desired overall variance of mixture normal distribution.

num_dist
Number of normal distributions.

means
Vector of means for each normal distribution. Must equal num_dist.

equalWeight
Should equal weights be used, only TRUE is currently supported.

Details

This function can be used to generate the inputs for the rbimod variances when a specific variance is desired. Especially useful when attempting to simulate a mixture normal/bimodal distribution.

Examples

# \donttest{
# calculating variance to be 2.5 with 2 distributions
desireVar(2.5, 2, means = c(-1, 1), equalWeight = TRUE)

#> [1] 1.5

# }