Simulate continuous variables

Function that simulates continuous variables. Any distribution function in R is supported.

sim_continuous2(
  n,
  dist = "rnorm",
  var_level = 1,
  variance = NULL,
  ther_sim = FALSE,
  ther_val = NULL,
  ...
)

Arguments

n
A list of sample sizes.

dist
A distribution function. This argument takes a quoted R distribution function (e.g. 'rnorm'). Default is 'rnorm'.

var_level
The level the variable should be simulated at. This can either be 1, 2, or 3 specifying a level 1, level 2, or level 3 variable respectively.

variance
The variance for random effect simulation.

ther_sim
A TRUE/FALSE flag indicating whether the error simulation function should be simulated, that is should the mean and standard deviation used for standardization be simulated.

ther_val
A vector of 2 that should include the theoretical mean and standard deviation of the generating function.

...
Additional parameters to pass to the dist_fun argument.