Function to simulate power.

Input simulation conditions and which term to compute power for, export reported power.

sim_pow_single(
  fixed,
  fixed_param,
  cov_param,
  n,
  error_var,
  with_err_gen,
  arima = FALSE,
  data_str,
  cor_vars = NULL,
  fact_vars = list(NULL),
  lvl1_err_params = NULL,
  arima_mod = list(NULL),
  contrasts = NULL,
  homogeneity = TRUE,
  heterogeneity_var = NULL,
  knot_args = list(NULL),
  missing = FALSE,
  missing_args = list(NULL),
  pow_param = NULL,
  alpha,
  pow_dist = c("z", "t"),
  pow_tail = c(1, 2),
  lm_fit_mod = NULL,
  general_mod = NULL,
  general_extract = NULL,
  ...
)

Arguments

fixed
One sided formula for fixed effects in the simulation. To suppress intercept add -1 to formula.

fixed_param
Fixed effect parameter values (i.e. beta weights). Must be same length as fixed.

cov_param
List of arguments to pass to the continuous generating function, must be the same order as the variables specified in fixed. This list does not include intercept, time, factors, or interactions. Required arguments include:

dist_fun: This is a quoted R distribution function.

var_type: This is the level of variable to generate. Must be 'single'. Must be same order as fixed formula above.

Optional arguments to the distribution functions are in a nested list, see the examples or vignettes for example code.

n
Cluster sample size.

error_var
Scalar of error variance.

with_err_gen
Simulated within cluster error distribution. Must be a quoted 'r' distribution function.

arima
TRUE/FALSE flag indicating whether residuals should be correlated. If TRUE, must specify a valid model to pass to arima.sim via the arima_mod argument. See arima.sim for examples.

data_str
Type of data. Must be "cross", "long", or "single".

cor_vars
A vector of correlations between variables.

fact_vars
A nested list of factor, categorical, or ordinal variable specification, each list must include:

numlevels: Number of levels for ordinal or factor variables.

var_type: Must be 'single'.

Optional arguments include:

replace

prob

value.labels

See also sample for use of these optional arguments.

lvl1_err_params
Additional parameters passed as a list on to the level one error generating function

arima_mod
A list indicating the ARIMA model to pass to arima.sim. See arima.sim for examples.

contrasts
An optional list that specifies the contrasts to be used for factor variables (i.e. those variables with .f or .c). See contrasts for more detail.

homogeneity
Either TRUE (default) indicating homogeneity of variance assumption is assumed or FALSE to indicate desire to generate heterogeneity of variance.

heterogeneity_var
Variable name as a character string to use for heterogeneity of variance simulation.

knot_args
A nested list of named knot arguments. See sim_knot for more details. Arguments must include:

var

knot_locations

missing
TRUE/FALSE flag indicating whether missing data should be simulated.

missing_args
Additional missing arguments to pass to the missing_data function. See missing_data for examples.

pow_param
Name of variable to calculate power for, must be a name from fixed.

alpha
What should the per test alpha rate be used for the hypothesis testing.

pow_dist
Which distribution should be used when testing hypothesis test, z or t?

pow_tail
One-tailed or two-tailed test?

lm_fit_mod
Valid lm syntax to be used for model fitting.

general_mod
Valid model syntax. This syntax can be from any R package. By default, broom is used to extract model result information. Note, package must be defined or loaded prior to running the sim_pow function.

general_extract
A valid function to extract model results if general_mod argument is used. This argument is primarily used if extracting model results is not possibly using the broom package. If this is left NULL (default), broom is used to collect model results.

...
Additional specification needed to pass to the random generating function defined by with_err_gen.

Details

Power function to compute power for a regression term for simple regression models. This function would need to be replicated to make any statement about power. Use sim_pow as a convenient wrapper for this.

See also