Master continuous simulation function.

Takes simulation parameters as inputs and returns simulated data.

sim_reg(
  fixed,
  random,
  random3,
  fixed_param,
  random_param = list(),
  random_param3 = list(),
  cov_param,
  k,
  n,
  p,
  error_var,
  with_err_gen,
  arima = FALSE,
  data_str,
  cor_vars = NULL,
  fact_vars = list(NULL),
  unbal = list(level2 = FALSE, level3 = FALSE),
  unbal_design = list(level2 = NULL, level3 = NULL),
  lvl1_err_params = NULL,
  arima_mod = list(NULL),
  contrasts = NULL,
  homogeneity = TRUE,
  heterogeneity_var = NULL,
  cross_class_params = NULL,
  knot_args = list(NULL),
  ...
)

Arguments

fixed
One sided formula for fixed effects in the simulation. To suppress intercept add -1 to formula.

random
One sided formula for random effects in the simulation. Must be a subset of fixed.

random3
One sided formula for random effects at third level in the simulation. Must be a subset of fixed (and likely of random).

fixed_param
Fixed effect parameter values (i.e. beta weights). Must be same length as fixed.

random_param
A list of named elements that must contain:

random_var = variance of random parameters,

rand_gen = Name of simulation function for random effects.

Optional elements are:

ther: Theorectial mean and variance from rand_gen,

ther_sim: Simulate mean/variance for standardization purposes,

cor_vars: Correlation between random effects,

...: Additional parameters needed for rand_gen function.

random_param3
A list of named elements that must contain:

random_var = variance of random parameters,

rand_gen = Name of simulation function for random effects.

Optional elements are:

ther: Theorectial mean and variance from rand_gen,

ther_sim: Simulate mean/variance for standardization purposes,

cor_vars: Correlation between random effects,

...: Additional parameters needed for rand_gen function.

cov_param
List of arguments to pass to the continuous generating function, must be the same order as the variables specified in fixed. This list does not include intercept, time, factors, or interactions. Required arguments include:

dist_fun: This is a quoted R distribution function.

var_type: This is the level of variable to generate. Must be either 'single', 'level1', 'level2', or 'level3'. Must be same order as fixed formula above.

Optional arguments to the distribution functions are in a nested list, see the examples or vignettes for example code.

k
Number of third level clusters.

n
Cluster sample size.

p
Within cluster sample size.

error_var
Scalar of error variance.

with_err_gen
Distribution function to pass on to the level one simulation of errors.

arima
TRUE/FALSE flag indicating whether residuals should be correlated. If TRUE, must specify a valid model to pass to arima.sim via the arima_mod argument. See arima.sim for examples.

data_str
Type of data. Must be "cross", "long", or "single".

cor_vars
A vector of correlations between variables.

fact_vars
A nested list of factor, categorical, or ordinal variable specification, each list must include:

numlevels = Number of levels for ordinal or factor variables.

var_type = Must be 'single', 'level1', 'level2', or 'level3'.

Optional arguments include:

replace

prob

value.labels

See also sample for use of these optional arguments.

unbal
A named TRUE/FALSE list specifying whether unbalanced simulation design is desired. The named elements must be: "level2" or "level3" representing unbalanced simulation for level two and three respectively. Default is FALSE, indicating balanced sample sizes at both levels.

unbal_design
When unbal = TRUE, this specifies the design for unbalanced simulation in one of two ways. It can represent the minimum and maximum sample size within a cluster via a named list. This will be drawn from a random uniform distribution with min and max specified. Secondly, the actual sample sizes within each cluster can be specified. This takes the form of a vector that must have the same length as the level two or three sample size. These are specified as a named list in which level two sample size is controlled via "level2" and level three sample size is controlled via "level3".

lvl1_err_params
Additional parameters passed as a list on to the level one error generating function

arima_mod
A list indicating the ARIMA model to pass to arima.sim. See arima.sim for examples.

contrasts
An optional list that specifies the contrasts to be used for factor variables (i.e. those variables with .f or .c). See contrasts for more detail.

homogeneity
Either TRUE (default) indicating homogeneity of variance assumption is assumed or FALSE to indicate desire to generate heterogeneity of variance.

heterogeneity_var
Variable name as a character string to use for heterogeneity of variance simulation.

cross_class_params
A list of named parameters when cross classified data structures are desired. Must include the following arguments:

num_ids: The number of cross classified clusters. These are in addition to the typical cluster ids

random_param: This argument is a list of arguments passed to sim_rand_eff. These must include:

random_var: The variance of the cross classified random effect

rand_gen: The random generating function used to generate the cross classified random effect.

Optional elements are:

ther: Theorectial mean and variance from rand_gen,

ther_sim: Simulate mean/variance for standardization purposes,

cor_vars: Correlation between random effects,

...: Additional parameters needed for rand_gen function.

knot_args
A nested list of named knot arguments. See sim_knot for more details. Arguments must include:

var

knot_locations

...
Not currently used.

Details

Simulated data is useful for classroom demonstrations and to study the impacts of assumption violations on parameter estimates, statistical power, or empirical type I error rates.

This function allows researchers a flexible approach to simulate regression models, including single level models and cross sectional or longitudinal linear mixed models (aka. hierarchical linear models or multilevel models).

Examples

# generating parameters for single level regression
fixed <- ~1 + act + diff + numCourse + act:numCourse
fixed_param <- c(2, 4, 1, 3.5, 2)
cov_param <- list(dist_fun = c('rnorm', 'rnorm', 'rnorm'), 
   var_type = c("single", "single", "single"),
   opts = list(list(mean = 0, sd = 4), 
   list(mean = 0, sd = 3),
   list(mean = 0, sd = 3)))
n <- 150
error_var <- 3
with_err_gen <- 'rnorm'
temp_single <- sim_reg(fixed = fixed, fixed_param = fixed_param, 
   cov_param = cov_param, 
   n = n, error_var = error_var, with_err_gen = with_err_gen, 
   data_str = "single")

#> Warning: 'sim_reg' is no longer supported and will be removed with v 1.0
#> 
#>               See tidy_simulation vignette.

# Fitting regression to obtain parameter estimates
summary(lm(sim_data ~ 1 + act + diff + numCourse + act:numCourse, 
   data = temp_single))

#> 
#> Call:
#> lm(formula = sim_data ~ 1 + act + diff + numCourse + act:numCourse, 
#>     data = temp_single)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -4.2808 -1.0185 -0.0534  0.8892  5.4641 
#> 
#> Coefficients:
#>               Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)    1.94731    0.13977   13.93   <2e-16 ***
#> act            3.96356    0.03794  104.48   <2e-16 ***
#> diff           0.97549    0.04626   21.09   <2e-16 ***
#> numCourse      3.49988    0.04863   71.97   <2e-16 ***
#> act:numCourse  2.00024    0.01359  147.15   <2e-16 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 1.708 on 145 degrees of freedom
#> Multiple R-squared:  0.9965,	Adjusted R-squared:  0.9964 
#> F-statistic: 1.045e+04 on 4 and 145 DF,  p-value: < 2.2e-16
#> 

# Longitudinal linear mixed model example
fixed <- ~1 + time + diff + act + time:act
random <- ~1 + time + diff
fixed_param <- c(4, 2, 6, 2.3, 7)
random_param <- list(random_var = c(7, 4, 2), rand_gen = 'rnorm')
cov_param <- list(dist_fun = c('rnorm', 'rnorm'), 
  var_type = c("level1", "level2"),
  opts = list(list(mean = 0, sd = 1.5), 
  list(mean = 0, sd = 4)))
n <- 150
p <- 30
error_var <- 4
with_err_gen <- 'rnorm'
data_str <- "long"
temp_long <- sim_reg(fixed, random, random3 = NULL, fixed_param, 
   random_param, random_param3 = NULL,
   cov_param, k = NULL, n, p, error_var, with_err_gen, data_str = data_str)

#> Warning: 'sim_reg' is no longer supported and will be removed with v 1.0
#> 
#>               See tidy_simulation vignette.

## fitting lmer model
library(lme4)

#> Loading required package: Matrix

lmer(sim_data ~ 1 + time + diff + act + time:act + 
  (1 + time + diff | clustID), 
  data = temp_long)

#> Linear mixed model fit by REML ['lmerMod']
#> Formula: sim_data ~ 1 + time + diff + act + time:act + (1 + time + diff |  
#>     clustID)
#>    Data: temp_long
#> REML criterion at convergence: 21339.01
#> Random effects:
#>  Groups   Name        Std.Dev. Corr       
#>  clustID  (Intercept) 2.916               
#>           time        2.126    -0.02      
#>           diff        1.510    -0.16  0.03
#>  Residual             1.996               
#> Number of obs: 4500, groups:  clustID, 150
#> Fixed Effects:
#> (Intercept)         time         diff          act     time:act  
#>       3.725        1.866        5.853        2.228        7.042  

# Three level example
fixed <- ~1 + time + diff + act + actClust + time:act
random <- ~1 + time + diff
random3 <- ~ 1 + time
fixed_param <- c(4, 2, 6, 2.3, 7, 0)
random_param <- list(random_var = c(7, 4, 2), rand_gen = 'rnorm')
random_param3 <- list(random_var = c(4, 2), rand_gen = 'rnorm')
cov_param <- list(dist_fun = c('rnorm', 'rnorm', 'rnorm'), 
     var_type = c("level1", "level2", "level3"),
     opts = list(list(mean = 0, sd = 1.5),
     list(mean = 0, sd = 4),
     list(mean = 0, sd = 2)))
k <- 10
n <- 15
p <- 10
error_var <- 4
with_err_gen <- 'rnorm'
data_str <- "long"
temp_three <- sim_reg(fixed, random, random3, fixed_param, random_param, 
random_param3, cov_param, k,n, p, error_var, with_err_gen, 
   data_str = data_str)

#> Warning: 'sim_reg' is no longer supported and will be removed with v 1.0
#> 
#>               See tidy_simulation vignette.

library(lme4)
lmer(sim_data ~ 1 + time + diff + act + actClust + time:act + 
   (1 + time + diff | clustID) +  
   (1 | clust3ID), data = temp_three)

#> Linear mixed model fit by REML ['lmerMod']
#> Formula: sim_data ~ 1 + time + diff + act + actClust + time:act + (1 +  
#>     time + diff | clustID) + (1 | clust3ID)
#>    Data: temp_three
#> REML criterion at convergence: 7894.225
#> Random effects:
#>  Groups   Name        Std.Dev. Corr       
#>  clustID  (Intercept) 3.022               
#>           time        2.125    -0.12      
#>           diff        1.416    -0.10  0.06
#>  clust3ID (Intercept) 1.913               
#>  Residual             2.030               
#> Number of obs: 1500, groups:  clustID, 150; clust3ID, 10
#> Fixed Effects:
#> (Intercept)         time         diff          act     actClust     time:act  
#>    2.477047     0.988698     6.084495     2.353936     7.071122    -0.008408