simglm: A package to simulate and perform power by simulation for models based on the generalized linear model.

The simglm package provides two categories of important functions: simulation functions (sim_reg and sim_glm) and power functions (sim_pow and sim_pow_glm). #'

This function is most useful to pass to replicate_simulation. The function attempts to determine automatically which aspects to add to the simulation/power generation based on the elements found in the sim_args argument.

simglm(sim_args)

Arguments

sim_args
A named list with special model formula syntax. See details and examples for more information. The named list may contain the following:

fixed: This is the fixed portion of the model (i.e. covariates)

random: This is the random portion of the model (i.e. random effects)

error: This is the error (i.e. residual term).